+++ changelog.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ## 0.0.2
 - Feature: Added Compass Support for Liferay 6.1 and 6.2
 - Feature: Added Liferay specific contexts for Live Templates 
+- Feature: Compare JSPs with original Liferay file (JSP hooks or web fragments)
 
 ## 0.0.1
 - Initial Version